+++ CHANGES.rst
@@ -31,6 +31,8 @@ Unreleased
    fetch the name of an unloaded Subreddit, use ``str(sub_instance)``, or
    ``unicode(sub_instance)``.
  * **[CHANGE]** Removed :class:`praw.Config` instance attribute ``is_reddit``.
+ * **[CHANGE]** :meth:`~praw.__init__.BaseReddit.evict` now returns the number
+   of items evicted.
  * **[FEATURE]** Added :meth:`get_message` to fetch a single Message object
    by its ID.
  * **[FEATURE]** Added :meth:`get_sticky` to get a Subreddit's sticky post.